fixup! refactor(providers/fab): move fab provider to new structure
