+++ providers/fab/tests/provider_tests/fab/auth_manager/api_endpoints/api_connexion_utils.py
@@ -25,6 +25,18 @@ with ignore_provider_compatibility_error("2.9.0+", __file__):
 
 
 @contextmanager
+def create_test_client(app, user_name, role_name, permissions):
+    """
+    Helper function to create a client with a temporary user which will be deleted once done
+    """
+    client = app.test_client()
+    with create_user_scope(app, username=user_name, role_name=role_name, permissions=permissions) as _:
+        resp = client.post("/login/", data={"username": user_name, "password": user_name})
+        assert resp.status_code == 302
+        yield client
+
+
+@contextmanager
 def create_user_scope(app, username, **kwargs):
     """
     Helper function designed to be used with pytest fixture mainly.
