+++ tests/www/views/test_views_home.py
@@ -27,8 +27,8 @@ from airflow.security import permissions
 from airflow.utils.state import State
 from airflow.www.utils import UIAlert
 from airflow.www.views import FILTER_LASTRUN_COOKIE, FILTER_STATUS_COOKIE, FILTER_TAGS_COOKIE
+from providers.fab.tests.provider_tests.fab.auth_manager.api_endpoints.api_connexion_utils import create_user
 
-from providers.tests.legacy_fab.api_connexion_utils import create_user
 from tests_common.test_utils.db import clear_db_dags, clear_db_import_errors, clear_db_serialized_dags
 from tests_common.test_utils.permissions import _resource_name
 from tests_common.test_utils.www import (
